Moved position calculations to setup-#11
Position, layer and position from left to right data is now calculated for each node in setup so this information can be used to render the tree. Additionally all nodes are stored in an array so there is the option of iteration through all nodes or traversing the tree. New script in draw is not identical to the previous in terms of spacing but is very similar and can be easily adjusted using variables at the top of the script.
Pretty sure this is clearly commented.
(Position from left to right is position as if the layer was full)